make service stateless
