+++ app/store/services/database.js
@@ -1,16 +1,13 @@
-import Config from "react-native-config";
 import { database } from "./helpers/database";
 
-export const authenticate = () => {
-  const configuration = {
-    apiKey: Config.FIREBASE_APIKEY,
-    authDomain: Config.FIREBASE_AUTHDOMAIN,
-    databaseURL: Config.FIREBASE_DATABASEURL,
-    storageBucket: Config.FIREBASE_STORAGEBUCKET
-  };
-
-  return database.connect(configuration).catch(error => ({ error }));
-};
+export const authenticate = (
+  configuration = {
+    apiKey: "",
+    authDomain: "",
+    databaseURL: "",
+    storageBucket: ""
+  }
+) => database.connect(configuration).catch(error => ({ error }));
 
 export const disconnect = () => {
   database.disconnect();
